refactor: Idの型をbyteに変更

+++ EnglishApp.Domain/Entities/UserGenderEntity.cs
@@ -1,7 +1,7 @@
 ﻿namespace EnglishApp.Domain.Entities;
 
-public sealed class UserGenderEntity(int id, string name)
+public sealed class UserGenderEntity(byte id, string name)
 {
-    public int Id { get; } = id;
+    public byte Id { get; } = id;
     public string Name { get; } = name;
 }
